refactor: migrate from BLoC to Cubit and modularize landing page components
- Replace PortfolioBloc with PortfolioCubit using separate state file pattern
- Extract landing page sections into reusable stateless widgets:
  * HeroSection, AboutSection, ExperienceSection
  * ProjectsSection, SkillsSection, ContactSection
- Remove redundant page files (about, contact, experience, home, projects, skills)
- Clean up navigation and app structure
- Fix particle animation positioning in AnimatedBackground
- Update VS Code launch configurations for web development
- Improve code organization and maintainability

Files changed: 25 files, +1195 insertions, -2925 deletions
