--- lib/pages/about_page.dart
@@ -1,310 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:url_launcher/url_launcher.dart';
-import 'package:auto_size_text/auto_size_text.dart';
-import '../blocs/portfolio_bloc.dart';
-import '../widgets/app_bar_widget.dart';
-import '../widgets/navigation_drawer.dart';
-import '../utils/app_theme.dart';
-
-class AboutPage extends StatelessWidget {
-  const AboutPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: const AppBarWidget(title: 'About Me'),
-      drawer: const AppNavigationDrawer(),
-      body: BlocBuilder<PortfolioBloc, PortfolioState>(
-        builder: (context, state) {
-          if (state is PortfolioLoading) {
-            return const Center(
-              child: CircularProgressIndicator(
-                color: AppTheme.primaryColor,
-              ),
-            );
-          }
-
-          if (state is PortfolioError) {
-            return Center(
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  const Icon(
-                    Icons.error_outline,
-                    size: 64,
-                    color: Colors.red,
-                  ),
-                  const SizedBox(height: 16),
-                  AutoSizeText(
-                    'Error: ${state.message}',
-                    style: Theme.of(context).textTheme.titleMedium,
-                  ),
-                ],
-              ),
-            );
-          }
-
-          if (state is PortfolioLoaded) {
-            return SingleChildScrollView(
-              padding: const EdgeInsets.all(24),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  // Profile Section
-                  _buildProfileSection(context, state.personalInfo),
-                  const SizedBox(height: 32),
-
-                  // Contact Information
-                  _buildContactSection(context, state.personalInfo),
-                  const SizedBox(height: 32),
-
-                  // Education
-                  _buildEducationSection(context, state.education),
-                ],
-              ),
-            );
-          }
-
-          return const Center(
-            child: AutoSizeText('No data available'),
-          );
-        },
-      ),
-    );
-  }
-
-  Widget _buildProfileSection(BuildContext context, personalInfo) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                const CircleAvatar(
-                  radius: 40,
-                  backgroundColor: AppTheme.primaryColor,
-                  child: Icon(
-                    Icons.person,
-                    size: 50,
-                    color: Colors.white,
-                  ),
-                ),
-                const SizedBox(width: 16),
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      AutoSizeText(
-                        personalInfo.name,
-                        style: Theme.of(context).textTheme.headlineMedium,
-                      ),
-                      const SizedBox(height: 4),
-                      AutoSizeText(
-                        personalInfo.title,
-                        style:
-                            Theme.of(context).textTheme.titleMedium?.copyWith(
-                                  color: AppTheme.primaryColor,
-                                  fontWeight: FontWeight.w600,
-                                ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 24),
-            AutoSizeText(
-              'Profile',
-              style: Theme.of(context).textTheme.titleLarge,
-            ),
-            const SizedBox(height: 8),
-            AutoSizeText(
-              personalInfo.profile,
-              style: Theme.of(context).textTheme.bodyLarge,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildContactSection(BuildContext context, personalInfo) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            AutoSizeText(
-              'Contact Information',
-              style: Theme.of(context).textTheme.titleLarge,
-            ),
-            const SizedBox(height: 16),
-            _buildContactItem(
-              context,
-              Icons.email,
-              'Email',
-              personalInfo.email,
-              () => _launchUrl('mailto:${personalInfo.email}'),
-            ),
-            _buildContactItem(
-              context,
-              Icons.location_on,
-              'Location',
-              personalInfo.home,
-              null,
-            ),
-            _buildContactItem(
-              context,
-              Icons.link,
-              'LinkedIn',
-              personalInfo.linkedin,
-              () => _launchUrl('https://${personalInfo.linkedin}'),
-            ),
-            _buildContactItem(
-              context,
-              Icons.code,
-              'GitHub',
-              personalInfo.github,
-              () => _launchUrl(personalInfo.github),
-            ),
-            ...personalInfo.phoneNumbers.map((phone) => _buildContactItem(
-                  context,
-                  Icons.phone,
-                  'Phone',
-                  phone,
-                  () => _launchUrl('tel:$phone'),
-                )),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildContactItem(
-    BuildContext context,
-    IconData icon,
-    String label,
-    String value,
-    VoidCallback? onTap,
-  ) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 8),
-      child: Row(
-        children: [
-          Icon(
-            icon,
-            color: AppTheme.primaryColor,
-            size: 20,
-          ),
-          const SizedBox(width: 12),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                AutoSizeText(
-                  label,
-                  style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                        color: AppTheme.textSecondaryColor,
-                      ),
-                ),
-                const SizedBox(height: 2),
-                GestureDetector(
-                  onTap: onTap,
-                  child: AutoSizeText(
-                    value,
-                    style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                          color: onTap != null
-                              ? AppTheme.primaryColor
-                              : AppTheme.textPrimaryColor,
-                          decoration:
-                              onTap != null ? TextDecoration.underline : null,
-                        ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildEducationSection(BuildContext context, List education) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            AutoSizeText(
-              'Education',
-              style: Theme.of(context).textTheme.titleLarge,
-            ),
-            const SizedBox(height: 16),
-            ...education.map((edu) => _buildEducationItem(context, edu)),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildEducationItem(BuildContext context, edu) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 12),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              const Icon(
-                Icons.school,
-                color: AppTheme.primaryColor,
-                size: 20,
-              ),
-              const SizedBox(width: 12),
-              Expanded(
-                child: AutoSizeText(
-                  edu.institution,
-                  style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                        fontWeight: FontWeight.w600,
-                      ),
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 8),
-          Padding(
-            padding: const EdgeInsets.only(left: 32),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                AutoSizeText(
-                  '${edu.degree} in ${edu.field}',
-                  style: Theme.of(context).textTheme.bodyLarge,
-                ),
-                const SizedBox(height: 4),
-                AutoSizeText(
-                  '${edu.period} • ${edu.score}',
-                  style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                        color: AppTheme.textSecondaryColor,
-                      ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Future<void> _launchUrl(String url) async {
-    final uri = Uri.parse(url);
-    if (await canLaunchUrl(uri)) {
-      await launchUrl(uri);
-    }
-  }
-}
