--- lib/pages/contact_page.dart
@@ -1,321 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:url_launcher/url_launcher.dart';
-import 'package:auto_size_text/auto_size_text.dart';
-import '../blocs/portfolio_bloc.dart';
-import '../widgets/app_bar_widget.dart';
-import '../widgets/navigation_drawer.dart';
-import '../utils/app_theme.dart';
-
-class ContactPage extends StatefulWidget {
-  const ContactPage({super.key});
-
-  @override
-  State<ContactPage> createState() => _ContactPageState();
-}
-
-class _ContactPageState extends State<ContactPage> {
-  final _formKey = GlobalKey<FormState>();
-  final _nameController = TextEditingController();
-  final _emailController = TextEditingController();
-  final _messageController = TextEditingController();
-
-  @override
-  void dispose() {
-    _nameController.dispose();
-    _emailController.dispose();
-    _messageController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: const AppBarWidget(title: 'Contact'),
-      drawer: const AppNavigationDrawer(),
-      body: BlocBuilder<PortfolioBloc, PortfolioState>(
-        builder: (context, state) {
-          if (state is PortfolioLoading) {
-            return const Center(
-              child: CircularProgressIndicator(
-                color: AppTheme.primaryColor,
-              ),
-            );
-          }
-
-          if (state is PortfolioError) {
-            return Center(
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  const Icon(
-                    Icons.error_outline,
-                    size: 64,
-                    color: Colors.red,
-                  ),
-                  const SizedBox(height: 16),
-                  AutoSizeText(
-                    'Error: ${state.message}',
-                    style: Theme.of(context).textTheme.titleMedium,
-                  ),
-                ],
-              ),
-            );
-          }
-
-          if (state is PortfolioLoaded) {
-            return SingleChildScrollView(
-              padding: const EdgeInsets.all(24),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  AutoSizeText(
-                    'Get In Touch',
-                    style: Theme.of(context).textTheme.headlineLarge,
-                  ),
-                  const SizedBox(height: 8),
-                  AutoSizeText(
-                    'Feel free to reach out for collaborations or opportunities',
-                    style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                          color: AppTheme.textSecondaryColor,
-                        ),
-                  ),
-                  const SizedBox(height: 32),
-
-                  // Contact Information
-                  _buildContactInfo(context, state.personalInfo),
-                  const SizedBox(height: 32),
-
-                  // Contact Form
-                  _buildContactForm(context),
-                ],
-              ),
-            );
-          }
-
-          return const Center(
-            child: AutoSizeText('No data available'),
-          );
-        },
-      ),
-    );
-  }
-
-  Widget _buildContactInfo(BuildContext context, personalInfo) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            AutoSizeText(
-              'Contact Information',
-              style: Theme.of(context).textTheme.titleLarge,
-            ),
-            const SizedBox(height: 16),
-            _buildContactItem(
-              context,
-              Icons.email,
-              'Email',
-              personalInfo.email,
-              () => _launchUrl('mailto:${personalInfo.email}'),
-            ),
-            _buildContactItem(
-              context,
-              Icons.location_on,
-              'Location',
-              personalInfo.home,
-              null,
-            ),
-            _buildContactItem(
-              context,
-              Icons.link,
-              'LinkedIn',
-              personalInfo.linkedin,
-              () => _launchUrl('https://${personalInfo.linkedin}'),
-            ),
-            _buildContactItem(
-              context,
-              Icons.code,
-              'GitHub',
-              personalInfo.github,
-              () => _launchUrl(personalInfo.github),
-            ),
-            ...personalInfo.phoneNumbers.map((phone) => _buildContactItem(
-                  context,
-                  Icons.phone,
-                  'Phone',
-                  phone,
-                  () => _launchUrl('tel:$phone'),
-                )),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildContactItem(
-    BuildContext context,
-    IconData icon,
-    String label,
-    String value,
-    VoidCallback? onTap,
-  ) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 12),
-      child: Row(
-        children: [
-          Container(
-            width: 48,
-            height: 48,
-            decoration: BoxDecoration(
-              color: AppTheme.primaryColor.withValues(alpha: 0.1),
-              borderRadius: BorderRadius.circular(12),
-            ),
-            child: Icon(
-              icon,
-              color: AppTheme.primaryColor,
-              size: 24,
-            ),
-          ),
-          const SizedBox(width: 16),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                AutoSizeText(
-                  label,
-                  style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                        color: AppTheme.textSecondaryColor,
-                      ),
-                ),
-                const SizedBox(height: 4),
-                GestureDetector(
-                  onTap: onTap,
-                  child: AutoSizeText(
-                    value,
-                    style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                          color: onTap != null
-                              ? AppTheme.primaryColor
-                              : AppTheme.textPrimaryColor,
-                          decoration:
-                              onTap != null ? TextDecoration.underline : null,
-                        ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildContactForm(BuildContext context) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            AutoSizeText(
-              'Send Message',
-              style: Theme.of(context).textTheme.titleLarge,
-            ),
-            const SizedBox(height: 16),
-            Form(
-              key: _formKey,
-              child: Column(
-                children: [
-                  TextFormField(
-                    controller: _nameController,
-                    decoration: const InputDecoration(
-                      labelText: 'Name',
-                      prefixIcon: Icon(Icons.person),
-                    ),
-                    validator: (value) {
-                      if (value == null || value.isEmpty) {
-                        return 'Please enter your name';
-                      }
-                      return null;
-                    },
-                  ),
-                  const SizedBox(height: 16),
-                  TextFormField(
-                    controller: _emailController,
-                    decoration: const InputDecoration(
-                      labelText: 'Email',
-                      prefixIcon: Icon(Icons.email),
-                    ),
-                    keyboardType: TextInputType.emailAddress,
-                    validator: (value) {
-                      if (value == null || value.isEmpty) {
-                        return 'Please enter your email';
-                      }
-                      if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$')
-                          .hasMatch(value)) {
-                        return 'Please enter a valid email';
-                      }
-                      return null;
-                    },
-                  ),
-                  const SizedBox(height: 16),
-                  TextFormField(
-                    controller: _messageController,
-                    decoration: const InputDecoration(
-                      labelText: 'Message',
-                      prefixIcon: Icon(Icons.message),
-                      alignLabelWithHint: true,
-                    ),
-                    maxLines: 5,
-                    validator: (value) {
-                      if (value == null || value.isEmpty) {
-                        return 'Please enter your message';
-                      }
-                      return null;
-                    },
-                  ),
-                  const SizedBox(height: 24),
-                  SizedBox(
-                    width: double.infinity,
-                    child: ElevatedButton.icon(
-                      onPressed: _submitForm,
-                      icon: const Icon(Icons.send),
-                      label: const AutoSizeText('Send Message'),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  void _submitForm() {
-    if (_formKey.currentState!.validate()) {
-      // Here you would typically send the form data
-      // For now, we'll just show a success message
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: AutoSizeText('Message sent successfully!'),
-          backgroundColor: AppTheme.primaryColor,
-          behavior: SnackBarBehavior.floating,
-        ),
-      );
-
-      // Clear the form
-      _nameController.clear();
-      _emailController.clear();
-      _messageController.clear();
-    }
-  }
-
-  Future<void> _launchUrl(String url) async {
-    final uri = Uri.parse(url);
-    if (await canLaunchUrl(uri)) {
-      await launchUrl(uri);
-    }
-  }
-}
