+++ pubspec.yaml
@@ -41,16 +41,10 @@ dependencies:
   # UI Components
   cupertino_icons: ^1.0.8
   google_fonts: ^6.1.0
-  flutter_svg: ^2.0.9
   url_launcher: ^6.2.4
-  font_awesome_flutter: ^10.7.0
   auto_size_text: ^3.0.0
 
-  # Web Specific
-  universal_html: ^2.2.4
-
   # Utilities
-  intl: ^0.19.0
   json_annotation: ^4.8.1
 
 dev_dependencies:
@@ -74,10 +68,10 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  assets:
-    - assets/images/
-    - assets/icons/
-    - assets/data/
+  # assets:
+  #   - assets/images/
+  #   - assets/icons/
+  #   - assets/data/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
